requirements.txt
django~=4.0.3
djangorestframework~=3.13.1
psycopg2-binary~=2.9.3
python-dotenv==0.19.0
drf-yasg~=1.20.0
gunicorn~=20.1.0
django-silk~=4.3.0
